+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.4] - 2026-09-04
+
+### Added
+- `createParentStore`function that allows you to define a new entry in the `parentsStores` array.
+
+### Changed
+- `parentsStores` option accepts an array or an object.
+
+
 ## [v0.1.3] - 2026-09-01
 
 ### Fixed